instead of relying on content-location, store base uri in xhr
given the httpbis changes on content-location semantics (rfc7231
appendix b), we can't rely on the proxy to tell us what it wants
relative links to be resolved against.

instead, we're now always assuming that relative links in the proxy's
response are relative to the original URI (that's an assertion the
library user places when stating that an URI template is suitable as
crossSiteProxyTemplate).

in practice, there are two types of proxies:

* proxies that understand rdf: those join relative links in the original
  representation anyway, and typically can't express them as relative
  uris at all.

  the only issue i can think of are proxies that add a tripel
  `<$ORIGINAL_RESOURCE> proxy:munged_by </me> .`, which will result in
  wrong statements but relatively unlikely.

* plain reverse proxies that just set CORS headers: they might have set
  content-location earlier (or still to), but given the above changes,
  we can't use that any more.